# track-2/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "evaluation_request": {
            "$ref": "#/components/schemas/EvaluationRequest"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatHistoryTurn"
            },
            "maxItems": 6,
            "title": "History",
            "type": "array"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message",
          "dataset_id",
          "evaluation_request"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ChatToolReceipt"
            },
            "title": "Tool Calls",
            "type": "array"
          },
          "usage": {
            "$ref": "#/components/schemas/ChatUsage"
          }
        },
        "required": [
          "meta",
          "answer",
          "model",
          "usage",
          "tool_calls"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
...
      "ClaimsRequest": {
        "additionalProperties": false,
        "properties": {
          "evaluation_request": {
            "$ref": "#/components/schemas/EvaluationRequest"
          },
          "team": {
            "title": "Team",
            "type": "string"
          }
        },
        "required": [
          "team",
          "evaluation_request"
        ],
        "title": "ClaimsRequest",
        "type": "object"
      },
      "ClaimsResponse": {
        "additionalProperties": false,
        "properties": {
          "claims": {
            "additionalProperties": true,
            "title": "Claims",
            "type": "object"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "meta",
          "claims"
        ],
        "title": "ClaimsResponse",
        "type": "object"
      },
      "CpuMigrationParameters": {
        "additionalProperties": false,
        "properties": {
          "added_elapsed_hours_per_job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Bounds"
              },
              {
                "type": "null"
              }
            ]
          },
          "additional_cpu_core_hours_per_gpu_hour": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Bounds"
              },
              {
                "type": "null"
              }
            ]
          },
          "recoverable_fraction": {
            "$ref": "#/components/schemas/Bounds"
          },
          "rerun_gpu_hours_per_candidate_gpu_hour": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Bounds"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "recoverable_fraction",
          "additional_cpu_core_hours_per_gpu_hour",
          "rerun_gpu_hours_per_candidate_gpu_hour",
          "added_elapsed_hours_per_job"
        ],
        "title": "CpuMigrationParameters",
        "type": "object"
      },
...
      "DecisionConfig": {
        "additionalProperties": false,
        "properties": {
          "action_ids": {
            "items": {
              "enum": [
                "cpu_migration",
                "idle_session_reclaim"
              ],
              "type": "string"
            },
            "title": "Action Ids",
            "type": "array"
          },
          "allocation_order": {
            "items": {
              "enum": [
                "cpu_migration",
                "idle_session_reclaim"
              ],
              "type": "string"
            },
            "title": "Allocation Order",
            "type": "array"
          },
          "default_request": {
            "$ref": "#/components/schemas/EvaluationRequest"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "meta",
          "default_request",
          "action_ids",
          "allocation_order"
        ],
        "title": "DecisionConfig",
        "type": "object"
      },
...
      "ErrorInfo": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorInfo",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Evaluation": {
        "additionalProperties": false,
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionEvaluation"
            },
            "title": "Actions",
            "type": "array"
          },
          "audit_status": {
            "enum": [
              "not_run",
              "running",
              "completed",
              "partial",
              "failed"
            ],
            "title": "Audit Status",
            "type": "string"
          },
          "baseline": {
            "$ref": "#/components/schemas/Baseline"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "portfolio": {
            "$ref": "#/components/schemas/Portfolio"
          },
          "price_book_version": {
            "title": "Price Book Version",
            "type": "string"
          },
          "ranking_basis": {
            "title": "Ranking Basis",
            "type": "string"
          },
          "request": {
            "$ref": "#/components/schemas/EvaluationRequest"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "meta",
          "request",
          "price_book_version",
          "baseline",
          "actions",
          "portfolio",
          "ranking_basis",
          "audit_status",
          "warnings"
        ],
        "title": "Evaluation",
        "type": "object"
      },
      "EvaluationRequest": {
        "additionalProperties": false,
        "properties": {
          "cpu_migration": {
            "$ref": "#/components/schemas/CpuMigrationParameters"
          },
          "idle_session_reclaim": {
            "$ref": "#/components/schemas/IdleSessionParameters"
          },
          "pricing": {
            "$ref": "#/components/schemas/Pricing"
          },
          "selected_action_ids": {
            "items": {
              "enum": [
                "cpu_migration",
                "idle_session_reclaim"
              ],
              "type": "string"
            },
            "title": "Selected Action Ids",
            "type": "array"
          }
        },
        "required": [
          "selected_action_ids",
          "pricing",
          "cpu_migration",
          "idle_session_reclaim"
        ],
        "title": "EvaluationRequest",
        "type": "object"
      },
      "EvidencePage": {
        "additionalProperties": false,
        "properties": {
          "action_id": {
            "enum": [
              "cpu_migration",
              "idle_session_reclaim"
            ],
            "title": "Action Id",
            "type": "string"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/JobEvidenceRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "scope": {
            "enum": [
              "standalone",
              "marginal"
            ],
            "title": "Scope",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "meta",
          "action_id",
          "scope",
          "rows",
          "total",
          "offset",
          "limit",
          "next_offset"
        ],
        "title": "EvidencePage",
        "type": "object"
      },
...
      "FindingDetail": {
        "additionalProperties": false,
        "properties": {
          "causal_status": {
            "enum": [
              "available",
              "no_chain"
            ],
            "title": "Causal Status",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "detector_id": {
            "title": "Detector Id",
            "type": "string"
          },
          "evidence_origin": {
            "enum": [
              "real_telemetry",
              "synthetic_incident",
              "test_fixture"
            ],
            "title": "Evidence Origin",
            "type": "string"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "impact_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Kind"
          },
          "impact_scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Scope"
          },
          "job_ids": {
            "items": {
              "type": "string"
            },
            "title": "Job Ids",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "node_names": {
            "items": {
              "type": "string"
            },
            "title": "Node Names",
            "type": "array"
          },
          "reported_impact_gpu_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported Impact Gpu Hours"
          },
          "resource_ids": {
            "items": {
              "type": "string"
            },
            "title": "Resource Ids",
            "type": "array"
          },
          "root_cause_ids": {
            "items": {
              "type": "string"
            },
            "title": "Root Cause Ids",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "meta",
          "finding_id",
          "detector_id",
          "title",
          "description",
          "resource_ids",
          "root_cause_ids",
          "job_ids",
          "node_names",
          "evidence_origin",
          "impact_scope",
          "impact_kind",
          "reported_impact_gpu_hours",
          "causal_status",
          "method"
        ],
        "title": "FindingDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IdleSessionParameters": {
        "additionalProperties": false,
        "properties": {
          "added_elapsed_hours_per_job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Bounds"
              },
              {
                "type": "null"
              }
            ]
          },
          "recoverable_fraction": {
            "$ref": "#/components/schemas/Bounds"
          },
          "rerun_gpu_hours_per_candidate_gpu_hour": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Bounds"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "recoverable_fraction",
          "rerun_gpu_hours_per_candidate_gpu_hour",
          "added_elapsed_hours_per_job"
        ],
        "title": "IdleSessionParameters",
        "type": "object"
      },
...
      "InvestigationResponse": {
        "additionalProperties": false,
        "properties": {
          "investigation": {
            "$ref": "#/components/schemas/Investigation"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "meta",
          "investigation"
        ],
        "title": "InvestigationResponse",
        "type": "object"
      },
...
      "Pricing": {
        "additionalProperties": false,
        "properties": {
          "usd_per_cpu_core_hour": {
            "anyOf": [
              {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "integer"
                  }
                ],
                "ge": 0,
                "le": 1000000
              },
              {
                "type": "null"
              }
            ],
            "title": "Usd Per Cpu Core Hour"
          },
          "usd_per_gpu_hour": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              }
            ],
            "ge": 0,
            "le": 1000000,
            "title": "Usd Per Gpu Hour"
          }
        },
        "required": [
          "usd_per_gpu_hour"
        ],
        "title": "Pricing",
        "type": "object"
      },
...
    }
  },
  "info": {
...
    "title": "MGAI API (hackathon facsimile)",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/decision/actions/{action_id}/evidence": {
      "post": {
        "operationId": "evidence_v1_decision_actions__action_id__evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "title": "Action Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "standalone",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidencePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Evidence",
        "tags": [
          "Decision API"
        ]
      }
    },
    "/v1/decision/chat": {
      "post": {
        "operationId": "chat_v1_decision_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Chat",
        "tags": [
          "Decision API"
        ]
      }
    },
    "/v1/decision/claims": {
      "post": {
        "operationId": "claims_v1_decision_claims_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Claims",
        "tags": [
          "Decision API"
        ]
      }
    },
    "/v1/decision/config": {
      "get": {
        "operationId": "config_v1_decision_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Config",
        "tags": [
          "Decision API"
        ]
      }
    },
    "/v1/decision/evaluate": {
      "post": {
        "operationId": "evaluate_v1_decision_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evaluation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Evaluate",
        "tags": [
          "Decision API"
        ]
      }
    },
    "/v1/decision/findings/{finding_id}": {
      "get": {
        "operationId": "finding_v1_decision_findings__finding_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Finding",
        "tags": [
          "Decision API"
        ]
      }
    },
...
    "/v1/decision/investigations/node_recommendation_audit": {
      "get": {
        "operationId": "investigation_v1_decision_investigations_node_recommendation_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Investigation",
        "tags": [
          "Decision API"
        ]
      }
    },
    "/v1/decision/jobs/{job_id}": {
      "get": {
        "operationId": "job_v1_decision_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Job",
        "tags": [
          "Decision API"
        ]
      }
    },
...
  }
}